requirements.txt
discord.py[voice]
SQLAlchemy==1.2.8
PyNaCl
youtube-dl